+++ src-ext/extension.ts
@@ -5,17 +5,17 @@ import { createBoard, deleteBoard, openBoard } from './kanbanBoards';
 export function activate(context: vscode.ExtensionContext) {
   fetchGitUserName();
 
-  let disposable = vscode.commands.registerCommand('sharableKanbanBoards.create', () => {
+  let disposable = vscode.commands.registerCommand('shareableKanbanBoards.create', () => {
     createBoard(context);
   });
   context.subscriptions.push(disposable);
 
-  disposable = vscode.commands.registerCommand('sharableKanbanBoards.open', () => {
+  disposable = vscode.commands.registerCommand('shareableKanbanBoards.open', () => {
     openBoard(context);
   });
   context.subscriptions.push(disposable);
 
-  disposable = vscode.commands.registerCommand('sharableKanbanBoards.delete', () => {
+  disposable = vscode.commands.registerCommand('shareableKanbanBoards.delete', () => {
     deleteBoard();
   });
   context.subscriptions.push(disposable);